- Added new button to assemble, link, and run code without debugging all in one go  - Added new logic to fib.asm which PRINTS A NEWLINE because it was BOTHERING ME
