+++ tests/samples/fib.asm
@@ -5,6 +5,7 @@
      	.data
 fibs: 	.word   0 : 12        # "array" of 12 words to contain fib values
 size: 	.word  12             # size of "array"
+nl:     .asciiz "\n"          # newline character
 
 # Start of code.
       	.text
@@ -51,4 +52,10 @@ print5: 	lw   	$a0, 0($t0)      # load fibonacci number for syscall
       	addi 	$t0, $t0, 4      # increment address
      	addi 	$t1, $t1, -1     # decrement loop counter
       	bgtz 	$t1, print5      # repeat if not finished
+		
+		# Print newline character
+		li      $v0, SysPrintString
+		la      $a0, nl
+		syscall
+
       	jr   	$ra              # return
